chatbot/logic.py
import os
import google.generativeai as genai
from dotenv import load_dotenv
from typing import List, Dict

load_dotenv()

api_key = os.getenv("GEMINI_API_KEY")
if not api_key:
    raise ValueError("GEMINI_API_KEY no encontrada. Asegúrate de tenerla en tu archivo .env")
genai.configure(api_key=api_key)

generation_config = {
    "temperature": 0.7,
    "top_p": 1,
    "top_k": 1,
    "max_output_tokens": 2048,
}
safety_settings = [
    {"category": "HARM_CATEGORY_HARASSMENT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_HATE_SPEECH", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_SEXUALLY_EXPLICIT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
    {"category": "HARM_CATEGORY_DANGEROUS_CONTENT", "threshold": "BLOCK_MEDIUM_AND_ABOVE"},
]
model = genai.GenerativeModel('gemini-1.5-flash', generation_config=generation_config, safety_settings=safety_settings)

def generate_bot_response(history: List[Dict], new_message: str) -> str:
    """Genera una respuesta usando el LLM con el historial de conversación."""
    
    gemini_history = []
    for msg in history:

        role = "model" if msg["role"] == "bot" else "user"
        gemini_history.append({"role": role, "parts": [msg["message"]]})

    try:
        chat_session = model.start_chat(history=gemini_history)
        response = chat_session.send_message(new_message)
        return response.text
    except Exception as e:
        print(f"Error al llamar a la API de Gemini: {e}")
        return "Tuve un problema para procesar mi respuesta. Intentemos de nuevo."